enable policy

+++ app/controllers/users_controller.rb
@@ -6,9 +6,9 @@ class UsersController < AdminController
     respond_to do |f|
       f.html do
         users        = list_users_filter
+        authorize User
         @users_count = users.count
         @users       = Kaminari.paginate_array(users).page(params[:page]).per(20)
-        # authorize @users
       end
       f.json { render json: UsersDatatable.new(view_context)}
     end
@@ -16,12 +16,12 @@ class UsersController < AdminController
 
   def new
     @user = User.new(enable_research_log_in: true)
-    # authorize @user
+    authorize @user
   end
 
   def create
     @user = User.new(user_params)
-    # authorize @user
+    authorize @user
     if @user.save
       redirect_to @user, notice: t('.successfully_created')
     else
@@ -30,15 +30,15 @@ class UsersController < AdminController
   end
 
   def show
-    # authorize @user
+    authorize @user
   end
 
   def edit
-    # authorize @user
+    authorize @user
   end
 
   def update
-    # authorize @user
+    authorize @user
     if @user.update_attributes(user_params)
       redirect_to @user, notice: t('.successfully_updated')
     else
@@ -47,7 +47,7 @@ class UsersController < AdminController
   end
 
   def destroy
-    # authorize @user
+    authorize @user
     if @user.destroy
       redirect_to users_url, notice: t('.successfully_deleted')
     else
